Applications & Bug Fixes
This update brings staff sign up commands, bug fixes and some changes. More can be found in the 1.0.0+ Changelog file

+++ commands/support/report.js
@@ -17,6 +17,10 @@ class ReportCommand extends Commando.Command
 
     async run(message, args)
     {
+        if (message.guild === null){
+            message.reply(DMMessage)
+            return;
+        }
         message.delete();
         let words = args.split(' ');
         let ReportedUser = words[0];
