+++ commands/support/suggest.js
@@ -17,6 +17,10 @@ class SuggestCommand extends Commando.Command
 
     async run(message, args)
     {
+        if (message.guild === null){
+            message.reply(DMMessage)
+            return;
+        }
         message.delete();
         let words = args.split(' ');
         let suggestion = words.slice(0).join(' ');
